# Fuel_Costs_TP Expression speedup