The Fuel_Costs_TP Expression in the fuel_cost module is taking significant time to be constructed. By building and using a new Set, this process can be sped up.

For the same input set used in my last commit (23 lz, 720 tp, 190 proj, 32 tx lines) building the Fuel_Costs_TP Expression was taking 7,76 s and the whole load_inputs function was at 71,04 s.
With these changes, the Expression construction takes 1,64 s and the load_inputs function takes 68,63 s.
Model instantiation is sped up only slightly (because the Expression akes less time to build, but a new Set has to be constructed), but I believe it may become more significant in larger models, specially with more timepoints.