@josiahjohnston 2016-08-23
A 5% speedup isn't bad :)
You can commit as-is if you like.

It might be slightly faster/cleaner to avoid defining DISPATCHED_PROJECTS_FUEL_IN_TP in dispatch.py, because it is only used in one operation in fuel_cost.py. You could instead define a custom set in fuel_cost that isn't a model component. I posted one way of doing that in the [fuel_cost_speedup_discussion](https://github.com/switch-model/switch/tree/fuel_cost_speedup_discussion) branch. I don't have profiling scripts set up to easily compare performance or try alternative ways of tuning that code, but in theory, having a local set could have less overhead than adding a new Set model component in a core module.