@mfripp 2016-08-24
I agree with Josiah that it's nice to keep down the number of sets if possible. DISPATCHED_PROJECTS_FUEL_IN_TP will probably only be used within the fuel_costs module, so it may not need to be added as a formal set (if it is added, maybe it should be renamed as TP_ACTIVE_PROJECTS_FUELS). On the other hand, the code in https://github.com/switch-model/switch/tree/fuel_cost_speedup_discussion is not as efficient as it could be, because it filters through all the projects active in each timepoint, which will include a lot of non-fuel projects. 

Another option would be to use code like this to define Fuel_Costs_TP (in fuel_cost.py):

    def Fuel_Costs_TP_rule(m, t):
        if not hasattr(m, 'Fuel_Costs_TP_dict'):
            # cache all Fuel_Cost_TP values in a dictionary (created in one pass)
            m.Fuel_Costs_TP_dict = {t2: 0.0 for t2 in m.TIMEPOINTS}
            for (proj, t2, f) in m.PROJ_FUEL_DISPATCH_POINTS:
                if (m.proj_load_zone[proj], f, m.tp_period[t2]) in m.FUEL_AVAILABILITY:
                    m.Fuel_Costs_TP_dict[t2] += (
                        m.ProjFuelUseRate[proj, t2, f] 
                        * m.fuel_cost[m.proj_load_zone[proj], f, m.tp_period[t2]]
                    )
        # return a result from the dictionary;
        # this uses pop() so all memory used by the dictionary will eventually be released
        return m.Fuel_Costs_TP_dict.pop(t)
    mod.Fuel_Costs_TP = Expression(mod.TIMEPOINTS, rule=Fuel_Costs_TP_rule)

I think this code will be about as fast as you can get. It seems to work fine at least with run_tests.py and some other spot checks, but I don't have a slow model that uses fuel_cost.py, so I can't say how it affects performance.