@josiahjohnston 2016-08-24
That looks great!
Benjamin, would you mind checking the speed of Matthias's code snippet?