@bmaluenda 2016-08-24
I ran a medium example a couple of times and whilst the original code takes between 4.3 and 4.8 s to construct the Fuel_Costs_TP expression, this new snippet reduces that time to between 0.85 and 0.95 s. Outputs were identical, so everything is working as expected.

It is true that defining new sets is something that should be avoided. I will commit this new version and if everything looks ok, I'll merge it in.