@josiahjohnston 2016-08-24
The code you posted has an indent error. Other than that, it looks good to me. Please squash it to a single commit with a clean message when you push the changes. 

Thanks! :)